+++ packages/iceworks-client/src/pages/Project/components/Dependencies/index.js
